made the clear-all-button functional
